fixed and simplified setting the user profile and loading all the streams in the proper order and on the root app file
