+++ pages/index.vue
@@ -14,7 +14,7 @@ useHead({
 
 definePageMeta({
   layout: 'default',
-  middleware: ['check-auth-provider', 'check-auth'],
+  middleware: ['check-auth-provider'],
 })
 const loginSideBar = useLoginSideBar()
 const signupSideBar = useSignupSideBar()
