+++ utilities/helpers.ts
@@ -1,7 +1,7 @@
 import { format, formatDistanceToNowStrict } from 'date-fns'
 import { Filesystem, Directory, Encoding } from '@capacitor/filesystem'
 import { StatusBar, Style } from '@capacitor/status-bar';
-import { useFileSystem, useAppDirectory, useCurrentEpisode, useTextSizeOption, useIsApp } from '~/composables/states'
+import { useFileSystem, useAppDirectory, useCurrentEpisode, useTextSizeOption, useIsApp, useCurrentUser, useCurrentUserProfile, useLocalUserProfileDefault } from '~/composables/states'
 import { Preferences } from '@capacitor/preferences';
 import {
   NativeSettings,
@@ -9,9 +9,11 @@ import {
   IOSSettings,
 } from 'capacitor-native-settings'
 import { Browser } from '@capacitor/browser';
+
+import { updateAllLiveStreams } from '~/composables/data/liveStream'
+//import { useSupabaseClient } from '@nuxtjs/supabase'
+
 const directoryToSaveTo = Directory.External
-
-
 
 // format ISO timestamp to return only the time
 export function formatTime(date: any) {
@@ -435,4 +437,115 @@ export const shareAPI = async (content: object) => {
 export const convertTime = (val) => {
   const hhmmss = new Date(val * 1000).toISOString().substring(11, 19)
   return hhmmss.startsWith('00:') ? hhmmss.substring(3) : hhmmss
+}
+
+
+export const getAndSetUserProfile = async () => {
+  const currentUser = useCurrentUser()
+  const currentUserProfile = useCurrentUserProfile()
+  const localUserProfileDefault = useLocalUserProfileDefault()
+  const config = useRuntimeConfig()
+  const client = useSupabaseClient()
+  const user = await client.auth.getSession()
+
+  // function that gets a user profile
+  const getProfile = async () => {
+    const {
+      data,
+      error
+    } = await client
+      .from('profiles')
+      .select('*')
+      .eq('id', currentUser.value.id)
+      .single()
+    if (error) {
+      console.error(error)
+    } else if (data) {
+      if (data.initial) {
+        // if first time logging in with new profile
+        const lsSTRING = await Preferences.get({ key: 'localUserProfile' })
+        const ls = JSON.parse(lsSTRING.value)
+        data.initial = false
+        data.autodownload = ls.autodownload
+        data.default_live_stream = ls.default_live_stream.label
+        data.receive_general_notifications = ls.receive_general_notifications
+        data.dark_mode = ls.dark_mode
+        data.text_size = ls.text_size.label
+
+        // update supabase profile data
+        // set the supabase prferences with what is currently set in the local storage
+        await client
+          .from('profiles')
+          .update({
+            initial: false,
+            autodownload: ls.autodownload,
+            default_live_stream: ls.default_live_stream.label,
+            receive_general_notifications: ls.receive_general_notifications,
+            dark_mode: ls.dark_mode,
+            text_size: ls.text_size.label,
+          })
+          .match({ id: currentUser.value.id })
+      }
+
+      // set the current user profile state
+      currentUserProfile.value = data
+      updateAllLiveStreams()
+      setDisplaySettings(data)
+    }
+  }
+
+  // check local storage for the auth token
+  if (process.client) {
+    const supabaseAuthToken = await Preferences.get({ key: config.public.supabaseAuthTokenName })
+
+    if (supabaseAuthToken.value) {
+      currentUser.value = JSON.stringify(supabaseAuthToken.user)
+    }
+
+    // check supabase session for logged in user
+    if (user?.data?.session?.user) {
+      currentUser.value = user?.data?.session?.user
+    }
+
+
+    if (!currentUser.value) {
+      // initially set default user profile settings or use the local storage settings
+
+      // does local storage settings exist?
+      const isLocalUserProfile = await Preferences.get({ key: 'localUserProfile' })
+      if (!isLocalUserProfile.value) {
+        // no, set defaults from localUserProfileDefault state
+        const defaults = localUserProfileDefault.value
+
+        //get the system's current theme and apply it to the initial defaults              
+        defaults.dark_mode = detectSystemDarkMode()
+
+        const defaultsSTRING = JSON.stringify(defaults)
+        await Preferences.set({
+          key: 'localUserProfile',
+          value: defaultsSTRING
+        })
+        currentUserProfile.value = {}
+        currentUserProfile.value = localUserProfileDefault.value
+
+        updateAllLiveStreams()
+        //set display settings
+        setDisplaySettings(localUserProfileDefault.value)
+      } else {
+        // local storage is set, so set currentUserProfile to the local storage settings                
+        currentUserProfile.value = {}
+        currentUserProfile.value = JSON.parse(isLocalUserProfile.value)
+
+        updateAllLiveStreams()
+        //set display settings
+        setDisplaySettings(currentUserProfile.value)
+      }
+      //navigateTo('/home')
+
+    } else {
+      // if they are a user, get their profile data
+      getProfile()
+
+    }
+  }
 }
